+++ e2e/nextjs-routes.d.ts
@@ -7,9 +7,7 @@ declare module "nextjs-routes" {
     | { pathname: "/foos/[foo]"; query: Query<{ foo: string }> }
     | { pathname: "/"; query?: Query | undefined };
 
-  type Query<Params = {}> = Params & {
-    [key: string]: string;
-  };
+  type Query<Params = {}> = Params & { [key: string]: string | undefined };
 }
 
 declare module "next/link" {
@@ -17,7 +15,7 @@ declare module "next/link" {
   import type { LinkProps as NextLinkProps } from "next/dist/client/link";
   import type { PropsWithChildren, MouseEventHandler } from "react";
 
-  type RouteOrQuery = Route | Pick<Route, "query">;
+  type RouteOrQuery = Route | { query?: { [key: string]: string | undefined } };
 
   export interface LinkProps extends Omit<NextLinkProps, "href"> {
     href: RouteOrQuery;
@@ -51,7 +49,7 @@ declare module "next/router" {
     [K in Route as K["pathname"]]: Exclude<K["query"], undefined>;
   };
 
-  type RouteOrQuery = Route | Pick<Route, "query">;
+  type RouteOrQuery = Route | { query: { [key: string]: string | undefined } };
 
   export interface NextRouter<P extends Pathname = Pathname>
     extends Omit<Router, "push" | "replace"> {