+++ src/utils.ts
@@ -126,11 +126,9 @@ declare module "nextjs-routes" {
           return `{ pathname: '${route.pathname}'; query?: Query | undefined }`;
         }
       })
-      .join("\n  | ")}
+      .join("\n    | ")}
 
-  type Query<Params = {}> = Params & {
-    [key: string]: string;
-  }
+  type Query<Params = {}> = Params & { [key: string]: string | undefined };
 }
 
 declare module "next/link" {
@@ -138,7 +136,7 @@ declare module "next/link" {
   import type { LinkProps as NextLinkProps } from "next/dist/client/link";
   import type { PropsWithChildren, MouseEventHandler } from "react";
 
-  type RouteOrQuery = Route | Pick<Route, 'query'>;
+  type RouteOrQuery = Route | { query?: { [key: string]: string | undefined } };
 
   export interface LinkProps extends Omit<NextLinkProps, "href"> {
     href: RouteOrQuery;
@@ -172,13 +170,18 @@ declare module "next/router" {
     [K in Route as K["pathname"]]: Exclude<K["query"], undefined>;
   };
 
-  type RouteOrQuery = Route | Pick<Route, 'query'>;
+  type RouteOrQuery = Route | { query: { [key: string]: string | undefined } };
 
-  export interface NextRouter<P extends Pathname = Pathname> extends Omit<Router, "push" | "replace"> {
+  export interface NextRouter<P extends Pathname = Pathname>
+    extends Omit<Router, "push" | "replace"> {
     pathname: P;
-    route: P; 
+    route: P;
     query: QueryForPathname[P];
-    push(url: RouteOrQuery, as?: string, options?: TransitionOptions): Promise<boolean>;
+    push(
+      url: RouteOrQuery,
+      as?: string,
+      options?: TransitionOptions
+    ): Promise<boolean>;
     replace(
       url: RouteOrQuery,
       as?: string,